+++ lib/modules/i18n/i18n_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: i18n
-// Updated at: 2026-04-10 16:36:57
-// Commit: feat(i18n): add async i18n stream provider
+// Updated at: 2026-04-12 11:24:51
+// Commit: fix(i18n): fix memory leak by cancelling i18n subscription
 
 class i18nComponentState {
-  final String commitRef = '285';
-  final String lastMessage = 'feat(i18n): add async i18n stream provider';
-  final DateTime timestamp = DateTime.parse('2026-04-10T16:36:57+05:00');
+  final String commitRef = '330';
+  final String lastMessage = 'fix(i18n): fix memory leak by cancelling i18n subscription';
+  final DateTime timestamp = DateTime.parse('2026-04-12T11:24:51+05:00');
 }