+++ CHANGELOG.md
@@ -331,3 +331,4 @@
 - [2026-04-12 10:05] chore(theme): update dependencies and cleanup unused imports in theme
 - [2026-04-12 10:26] feat(storage): add async storage stream provider
 - [2026-04-12 11:12] test(navigation): implement widget integration tests for navigation
+- [2026-04-12 11:24] fix(i18n): fix memory leak by cancelling i18n subscription